hmr/requirements.txt
# python requirements
pip>=9.0
scipy
numpy
# opendr==0.77
matplotlib
scikit-image
deepdish>=0.3
opencv-python
absl-py
ipdb
